agent1_infrastructure/requirements.txt
# Agent 1 Infrastructure Foundation - Python Requirements

# Core dependencies
ansible>=2.10.0
PyYAML>=5.4.0

# SSH and networking
paramiko>=2.7.0
netaddr>=0.8.0

# Optional but recommended
jinja2>=2.11.0
requests>=2.25.0
